Create bowlingstats_display1.java
This is my first foray into Java Swing overall and the implementation of GUI for the bowling statistics program.  All it does right now is take some parameters of score data and print it out in rows.  A picture of the output can be seen in "output_v1_170123.png".  It is crude right now but looks alright for a first try.