Update bowlingstats_prelim1.java
Main change: I coded class bowlingstats_display1.java, which is implementation of the Java Swing GUI to display the scores more aesthetically.  It is very basic and crude but is my first time messing with Java GUI aside from random message boxes.  This version of the program now outputs both to console and to GUI format.  It creates a String array and an integer array and passes them to the bowlingstats_display1 class.  It will probably be the last of bowlingstats_prelim1, with a fully GUI interface starting with bowlingstats_prelim2.

Minor change: For the console output, I was having trouble lining up the "frame 1, frame 2" line with the other lines; my solution was crudely to add an opening space (previously lines 197, 205).  This was fixed by removing the after-post spacing ("    |" instead of "  |   "), which makes the code slightly cleaner and more compact.