Create bgame_v2.java
This bgame_v2 class is necessary for bowlingstats_prelim2.java to function.  (bowlingstats_prelim1 had bgame_v1 as an inner class, but at this point it is too unwieldy)
The changes compared to bgame_v1 are hard to describe, but I rewrote the code from scratch in far less time than it took for bgame_v1, and it felt smoother and more logical despite apparently taking a few more lines.  I also added some code for closed frames (so they can be shown as bold).