Fixed error in scoring algorithm
There was an error with games that finished with two strikes.  The criteria for the third strike was placed in the situation that the first ball was a strike and the second was not.  (this would not make sense anyway)  I moved it up a couple lines to the correct area- if the first and second balls are both strikes.
A test game was showing the third ball in a two-strike tenth frame as a null, but it now correctly displays a strike.