Fixed display spacing issue
The last version was having consistency problems with spacing.  When selecting different games, the size of the boxes and the position of the headers sometimes changed.  I found that it is because the headers are based on the position of the last text of the boxes, so if there was more text (e.g. tenth frame having bonus) or bold text, it shifted the JLabel header.  I think it is a flaw in the Box class but opted not to change that.
My solution was to encapsulate the JLabel header inside of a JPanel, then add that to the Box, instead of adding the JLabel to the Box.
I also temporarily removed the bold font for closed frames feature and replaced it with a color change for closed frames, only to add it back in and keep the color change.  Closed frames are now bold but also have a green background.