Update 3-14-17 (major)
Major update- changed interface to be able to view multiple games at once instead of only one game.   Games are now displayed vertically in a larger window that can be scrolled to view.  I copied a share of methods, but this required almost an overhaul of the code.

I changed the ActionListener from a helper method to a Lambda expression, partly to save space but also because I needed to force the master program to refresh its view.  Otherwise, when selecting a game(s), nothing displayed unless the window was resized.

The code is a big mess right now.  Some processes can probably be streamlined, and there are many old fields that I should delete.

There are problems with the sizing and alignment.
When the program is loaded, the game selection list is in the middle, instead of on the right with the game data space being open on the left.  I tried setting a minimum size for "mainDataDisplay" with no luck.  I succeeded in preventing single games from stretching vertically to the whole screen.
When more games are loaded, the game selection list is also moved down as the program attempts to center it vertically, instead of staying on the top.  For some reason the vertical alignment does not work on this, and I will need to investigate and fix that.  This is a real problem if more than about seven games are loaded, bumping the selection list off the screen.