Update 3-26-17
Major update- added early version of statistics.  The program now shows high, average, and low scores for the respective frame for all games displayed.  Example- five games are selected, and the best score on frame #4 for those five games is 74, the average is 54.2, and the worst score is 45.  This is shown for all ten frames.

Minor adjustments to layout and dimensions.  I did some experimentation with alignment/dimensions and tentatively concluded that a "PreferredSize" sets the initial size, while a "MaximumSize" is straightforward, and that left/top/etc. alignment cannot apply to text, only for a JPanel, as those methods are simply inherited from JComponent but cannot actually be used.