Update 4-02-17
Mostly minor changes to the code in accordance with fundamentally sound practices.  This includes removing fields that were not used (left over from when the program showed only one game at a time), eliminating methods that are used only once, renaming or removing local fields that hide instance fields, order and grouping, and adding a few comments.

A couple minor changes to the visual format.  Games in the game display are separated vertically (a buffer of 20 pixels).  Frame-by-frame averages are larger and easier to read.  (they no longer have "high"/"avg"/"low" labels, but these will be added back in on the side)  Program is resized to better fit the current display.